Use explicit docker image name throughout
This adds the 'ministryofjustice/' org. prefix
to the docker image name, and refers to the full
image name, including tag, when running the script

This will enable people to invoke the script
without having to build the docker image for 
themselves - when they invoke the script, docker
should automatically pull the right image from
docker hub.