Prevent error if tfstate object has no "modules"
Not every statefile in the S3 bucket that stores
our terraform states represents a namespace.

For those which are not namespaces, there may not
be a "modules" key in the hash contained in the
statefile.

This change causes the environments checker to 
skip any statefile which doesn't have a "modules"
key.

It might be desirable to have a more definitive
test for "is this the statefile of a namespace?", 
but this seems sufficient, for now.