Exclude infrastructure namespaces from "orphans"
If we define a namespace in the infrastructure
repo, then it shouldn't appear in the list of
"orphaned" namespaces (i.e. those without a folder in the environments repo).

This change adds an "InfrastructureNamespaceLister" class which outputs a list of the namespaces defined in the infrastructure repo terraform code.

All such namespaces are created using the "kubernetes_namespace" terraform resource, so the class greps the source code of the infrastructure repo for matching lines, and uses a regexp to extract the names of the namespaces being created. Any such namespaces are subtracted from the "orphans" list.

This approach is a bit brittle, but since all our terraform code will always be the output of `terraform fmt`, I think we can get away with it.

Anyway, it's better than having to update this project every time we define a new namespace in the infrastructure repo.